rag/requirements_api.txt
# RAG API 서버 실행에 필요한 패키지들
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0

# LangChain 관련 패키지들
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-chroma>=0.1.2
langchain-core>=0.1.0

# 기타 필요한 패키지들
python-dotenv>=1.0.0
python-multipart>=0.0.6
